requirements.txt
beautifulsoup4==4.12.3
certifi==2024.6.2
charset-normalizer==3.3.2
coloredlogs==15.0.1
contourpy==1.2.1
cycler==0.12.1
filelock==3.15.1
flatbuffers==24.3.25
fonttools==4.53.0
fsspec==2024.6.0
gdown==5.2.0
humanfriendly==10.0
idna==3.7
imageio==2.34.1
imgviz==1.7.5
isort==5.13.2
Jinja2==3.1.4
kiwisolver==1.4.5
labelme==5.5.0
lazy_loader==0.4
MarkupSafe==2.1.5
matplotlib==3.9.0
mpmath==1.3.0
natsort==8.4.0
networkx==3.3
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.19.3
nvidia-nvjitlink-cu12==12.5.40
nvidia-nvtx-cu12==12.1.105
onnxruntime==1.18.0
opencv-python==4.10.0.82
opencv-python-headless==4.10.0.82
packaging==24.1
pillow==10.3.0
protobuf==5.27.1
pyparsing==3.1.2
PyQt5==5.15.10
PyQt5-Qt5==5.15.2
PyQt5-sip==12.13.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
PyYAML==6.0.1
QtPy==2.4.1
readlif==0.6.5
requests==2.32.3
scikit-image==0.23.2
scipy==1.13.1
six==1.16.0
soupsieve==2.5
sympy==1.12.1
termcolor==2.4.0
tifffile==2024.5.22
torch==2.2.2
torchvision==0.17.2
tqdm==4.66.4
triton==2.2.0
typing_extensions==4.12.2
urllib3==2.2.1
